Rearrange the classes in packages. Introduce Dagger/Hilt and introduce repository layer.
