--- ui-app/src/main/kotlin/com/kavi/droid/color/palette/app/data/Quadruple.kt
@@ -1,8 +1,0 @@
-package com.kavi.droid.color.palette.app.data
-
-data class Quadruple<A, B, C, D>(
-    val first: A,
-    val second: B,
-    val third: C,
-    val fourth: D
-)
