+++ ui-app/src/main/kotlin/com/kavi/droid/color/palette/app/data/dto/Quadruple.kt
@@ -1,0 +1,8 @@
+package com.kavi.droid.color.palette.app.data.dto
+
+data class Quadruple<A, B, C, D>(
+    val first: A,
+    val second: B,
+    val third: C,
+    val fourth: D
+)
