+++ ui-app/src/main/kotlin/com/kavi/droid/color/palette/app/di/RepositoryModule.kt
@@ -1,0 +1,18 @@
+package com.kavi.droid.color.palette.app.di
+
+import com.kavi.droid.color.palette.app.data.repository.SettingsDataRepository
+import com.kavi.droid.color.palette.app.data.repository.SettingsLocalRepositoryImpl
+import dagger.Binds
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.components.ViewModelComponent
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(ViewModelComponent::class)
+abstract class RepositoryModule {
+    @Binds
+    abstract fun provideRepository(repository: SettingsLocalRepositoryImpl): SettingsDataRepository
+}
