--- ui-app/src/main/kotlin/com/kavi/droid/color/palette/app/model/PaletteType.kt
@@ -1,5 +1,0 @@
-package com.kavi.droid.color.palette.app.model
-
-enum class PaletteType {
-    ALPHA_PALETTE, SATURATION_PALETTE, LIGHTNESS_PALETTE, KV_PALETTE
-}
