--- ui-app/src/main/kotlin/com/kavi/droid/color/palette/app/model/TabItem.kt
@@ -1,6 +1,0 @@
-package com.kavi.droid.color.palette.app.model
-
-data class TabItem(
-    val name: String,
-    val icon: Int
-)
