--- ui-app/src/main/kotlin/com/kavi/droid/color/palette/app/theme/Color.kt
@@ -1,69 +1,0 @@
-package com.kavi.droid.color.palette.app.theme
-
-import androidx.compose.ui.graphics.Color
-
-val MatWhite = Color(0xAAFFFFFF)
-val MatBlack = Color(0xAA000000)
-val Mat100Red = Color(0xFFFFCDD2)
-val Mat100Rose = Color(0xFFF8BBD0)
-val Mat100LPurple = Color(0xFFE1BEE7)
-val Mat100DPurple = Color(0xFFD1C4E9)
-val Mat100DBlue = Color(0xFFC5CAE9)
-val Mat100LBlue = Color(0xFFBBDEFB)
-val Mat100LLBlue = Color(0xFFB3E5FC)
-val Mat100LCyan = Color(0xFFB2EBF2)
-val Mat100DCyan = Color(0xFFB2DFDB)
-val Mat100DGreen = Color(0xFFC8E6C9)
-val Mat100LGreen = Color(0xFFDCEDC8)
-val Mat100LLGreen = Color(0xFFF0F4C3)
-val Mat100Yellow = Color(0xFFFFF9C4)
-val Mat100Gold = Color(0xFFFFECB3)
-val Mat100Orange = Color(0xFFFFE0B2)
-
-val Mat500Red = Color(0xFFF44336)
-val Mat500Rose = Color(0xFFE91E63)
-val Mat500LPurple = Color(0xFF9C27B0)
-val Mat500DPurple = Color(0xFF673AB7)
-val Mat500DBlue = Color(0xFF3F51B5)
-val Mat500LBlue = Color(0xFF2196F3)
-val Mat500LLBlue = Color(0xFF03A9F4)
-val Mat500LCyan = Color(0xFF00BCD4)
-val Mat500DCyan = Color(0xFF009688)
-val Mat500DGreen = Color(0xFF4CAF50)
-val Mat500LGreen = Color(0xFF8BC34A)
-val Mat500LLGreen = Color(0xFFCDDC39)
-val Mat500Yellow = Color(0xFFFFEB3B)
-val Mat500Gold = Color(0xFFFFC107)
-val Mat500Orange = Color(0xFFFF9800)
-
-val Mat800Red = Color(0xFFC62828)
-val Mat800Rose = Color(0xFFAD1457)
-val Mat800LPurple = Color(0xFF6A1B9A)
-val Mat800DPurple = Color(0xFF4527A0)
-val Mat800DBlue = Color(0xFF283593)
-val Mat800LBlue = Color(0xFF1565C0)
-val Mat800LLBlue = Color(0xFF0277BD)
-val Mat800LCyan = Color(0xFF00838F)
-val Mat800DCyan = Color(0xFF00695C)
-val Mat800DGreen = Color(0xFF2E7D32)
-val Mat800LGreen = Color(0xFF558B2F)
-val Mat800LLGreen = Color(0xFF9E9D24)
-val Mat800Yellow = Color(0xFFF9A825)
-val Mat800Gold = Color(0xFFFF8F00)
-val Mat800Orange = Color(0xFFEF6C00)
-
-val Mat900Red = Color(0xFFB71C1C)
-val Mat900Rose = Color(0xFF880E4F)
-val Mat900LPurple = Color(0xFF4A148C)
-val Mat900DPurple = Color(0xFF311B92)
-val Mat900DBlue = Color(0xFF1A237E)
-val Mat900LBlue = Color(0xFF0D47A1)
-val Mat900LLBlue = Color(0xFF01579B)
-val Mat900LCyan = Color(0xFF006064)
-val Mat900DCyan = Color(0xFF004D40)
-val Mat900DGreen = Color(0xFF1B5E20)
-val Mat900LGreen = Color(0xFF33691E)
-val Mat900LLGreen = Color(0xFF827717)
-val Mat900Yellow = Color(0xFFF57F17)
-val Mat900Gold = Color(0xFFFF6F00)
-val Mat900Orange = Color(0xFFE65100)
