+++ ui-app/src/main/kotlin/com/kavi/droid/color/palette/app/ui/model/PaletteType.kt
@@ -1,0 +1,5 @@
+package com.kavi.droid.color.palette.app.ui.model
+
+enum class PaletteType {
+    ALPHA_PALETTE, SATURATION_PALETTE, LIGHTNESS_PALETTE, KV_PALETTE
+}
