+++ ui-app/src/main/kotlin/com/kavi/droid/color/palette/app/ui/model/TabItem.kt
@@ -1,0 +1,6 @@
+package com.kavi.droid.color.palette.app.ui.model
+
+data class TabItem(
+    val name: String,
+    val icon: Int
+)
